remove all image metadata more reliably
Recent Chrome is stricter when parsing AVIF images, and the fallback logic seems
to be broken. Our AVIF files had the "XMP" tag placed incorrectly in the ISO
container. XMP is metadata, which we tried to remove using exiftool already
anyway. The code was changed to have ImageMagick omit all metadata when writing
the AVIF in the first place.

Caution: This requires `rm -rf cache`, otherwise we'll keep re-uploading the
         cached versions of the AVIF files with the problem.