tools/mse: Fix MSE2.0 checksum
If the start address of a program begins relatively high in a page,
it makes a difference if the checksum seed (based on the start address
of the line and the start address of the program) is calculated from
only the high bits or if it's shifted late in the process.

It doesn't happen often (most listings begin at $xx00 oder $xx01), but
the August 1990 issue has a listing that triggers the bug.