Autogenerate 64er-mag style listings (#170)
* tools: Add petcat2checksummer utility

The python file can be imported or executed directly (as a crude form of
a demo on how to use the functions).

Given a petcat output and corresponding prg file, it creates a 64er
Magazine style listing.

* Specify BASIC version when tokenizing

The petcat(1) interface isn't _very_ clearly specified on that, so just
set both.

* factor out petcat-to-prg conversion

* Hook up petcat2checksummer with the site generator

* Hook up MSE generator

* Adapt MSE listings to use the new listing autogeneration

* Enable Checksummer support for a few listings

* tools/mse: Fix MSE2.0 checksum

If the start address of a program begins relatively high in a page,
it makes a difference if the checksum seed (based on the start address
of the line and the start address of the program) is calculated from
only the high bits or if it's shifted late in the process.

It doesn't happen often (most listings begin at $xx00 oder $xx01), but
the August 1990 issue has a listing that triggers the bug.