Refine listing layouts (#180)
* Move listing toggle into the box

... or at least make it look like it's in the box. It really is
an object on top of it, with the box leaving enough room for it to
work out.

* Add "copy" button to listings

If Javascript is available and we're in a secure context (so clipboard
access works), add a copy button to listings that always takes the
petcat sources.

* petcat2checksummer: Make text upper-case

Just like in the magazines. Originally upper-case characters now use the
shift modifier (underline)

* petcat2checksummer: collapse strides of spaces

64er 7/1990 has a listing (pg. 102/103) that shows that multiple spaces
are combined into {nSPACE} starting with n=2, and it is, in fact, more
readable.

* checksummer listings: Replace combining characters with HTML

The combining characters are a bit of an abuse (as these combinations
are unusual, to say the least) and so many fonts and/or font renderers
don't handle them properly.

Go for simple over/underlines instead.

* Rework Checksummer listing support

It now works directly from the .prg files (generated via petcat) and its
own token tables.

This eliminates a few of the less obvious hacks when working at the text
level, allows further clean-up and closer reproduction of the listings
as shown in the magazine.

* checksummer: Mirror space addition behavior around tokens

The magazine put in spaces around keywords for legibility. @mist64
reconstructed the rules, this reimplements them.

Spaces aren't added if there's already a space in the original file. I
_think_ this covers what the magazine did, too.

* checksummer: Simplify {SPACE} handling

* checksummer: Fix control code compression

Sometimes multiple compression opportunities overlapped and led to
incomplete results, for example:

{a}{a}{a} something {a}{a}

could have led to

{2a}{a} something {2a}

* generate.py: Default to Checksummer listing where available