Fixup listings (#249)
* generate: Guard petcat listings against HTML interpreted characters

When generating petcat/checksummer selectable listings, the petcat lines
were passed through 1:1, which led to trouble with "<".
To prevent such issues, entity-encode <, and also & because that's the
entity-header.

* checksummer: correctly identify spaces to coalesce

When deciding whether to add a space before or after a token, compare
the byte with the byte value of space, not with a string.

* checksummer: Map non-quoted characters, too

A petcat "_" is a left-arrow in a 64er listing, so use the conversion
table.

If there are odd bytes outside quotes, we're in trouble anyway. Might as
well try to print a useful C64-style input sequence for those.

* style.css: Use multi-column style for MSE listings automatically